fast_api_media_platform/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
jinja2
python-multipart
python-jose[cryptography]
passlib[bcrypt]
fastapi_mail
